==> 02_02_Preprocesado_definitivo.ipynb <==
Eliminamos los documentos con menos de dos palabras
<h3>Preprocesado</h3
<ol>
    <li>Tokenización y lematización</li>
    <li>Eliminar signos de puntuación, tokens numéricos y stopwords</li>
    <li>Eliminar acentos</li>
    <li>Eliminar tokens de 1 carácter</li>
    <li>Aplicar filter_extremes no_below=7 (Al menos aparezca en 7 documentos), no_above=0.7 (no aparezca en más del 70% de los documentos)</li>
</ol>
<h3>Análisis y Unificación de servicios: </h3>
<ol>
    <li>Análisis de servicios existentes con códigos diferentes</li>
    <li>Unificación de servicios para reducir su número</li>
</ol>    

import pandas as pd
import requests
import datetime
import re
import string
import collections
import spacy
import ast
import unicodedata
from gensim import corpora
from gensim.corpora import MmCorpus
from gensim.models import LdaModel
from collections import Counter
import copy
from sklearn.metrics import silhouette_score, adjusted_rand_score, normalized_mutual_info_score, davies_bouldin_score, calinski_harabasz_score
import bcubed
from sklearn.preprocessing import LabelEncoder
from sklearn.metrics.pairwise import cosine_similarity


df_original = pd.read_csv('qys_original.csv', sep=";")

def count_words(text):
    words = re.findall(r'\b\w+\b',text)
    return len(words)

df_word = df_original.copy()
df_word['texto'] = df_word['title'].astype(str) + ' ' + df_word['description'].astype(str)
df_word['count_word'] = df_word['texto'].apply(count_words)

df_word[['count_word','texto']]

#eliminamos los documentos con menos de dos palabras
df_word=df_word[df_word['count_word'] >= 3]

df_preprocesado = df_word.copy()

import spacy
import unicodedata

nlp = spacy.load("es_core_news_sm")

def preprocesar_textos(texto, generar_bigramas=False):
    # 1. Tokenización y lematización
    doc = nlp(texto.lower())  # Procesa el texto en minúsculas
    # 2. Eliminar signos de puntuación, tokens numéricos y stopwords
    tokens_sin_stopwords = [token.lemma_ for token in doc if not token.is_punct and token.is_alpha and not token.is_stop]

    # 3. Eliminar acentos
    tokens_sin_acentos = []
    for token in tokens_sin_stopwords:
        if 'ñ' in token or 'Ñ' in token:
            tokens_sin_acentos.append(token)
        else:
            normalized_token = unicodedata.normalize('NFKD', token)
            token_sin_acento = ''.join([c for c in normalized_token if unicodedata.category(c) != 'Mn'])
            tokens_sin_acentos.append(token_sin_acento)

    # 4. Eliminar tokens de 1 carácter
    tokens_filtrados = [token for token in tokens_sin_acentos if len(token) > 1]

    if generar_bigramas:
        # 5. Generar bigramas
        bigramas = [" ".join(bigrama) for bigrama in zip(tokens_filtrados, tokens_filtrados[1:])]
        return bigramas
    else:
        # 6. Devolver solo unigramas
        return tokens_filtrados


import nltk
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
def generar_bigramas(texto):
    """Genera bigramas a partir de un texto."""

    # Generación de bigramas
    bigramas = list(nltk.bigrams(texto))
    return bigramas


df_preprocesado = pd.read_csv('df_preprocesado.csv', sep=';')

df_preprocesado['texto_preprocesado'] = df_preprocesado['texto'].apply(preprocesar_texto)

df_preprocesado['bigramas_preprocesado'] = df_preprocesado['texto'].apply(lambda x: preprocesar_textos(x, generar_bigramas=True))

df_preprocesado['bi_nltk_preprocesado'] = df_preprocesado['texto_preprocesado'].apply(ast.literal_eval).apply(generar_bigramas)

df_preprocesado['texto_preprocesado']

df_preprocesado['bigramas_preprocesado']

df_preprocesado['bi_nltk_preprocesado']

 df_preprocesado.to_csv('df_preprocesado.csv', sep=';', index=False)

dictionary_spacy = corpora.Dictionary(df_preprocesado['texto_preprocesado'])
dictionary_spacy.save('dictionary_spacy.gensim')

dictionary_bi_spacy = corpora.Dictionary(df_preprocesado['bigramas_preprocesado'])
dictionary_bi_spacy.save('dictionary_bi_spacy.gensim')

len(dictionary_spacy)

len(dictionary_bi_spacy)

dictionary_bi_nltk = corpora.Dictionary(df_preprocesado['bi_nltk_preprocesado'].apply(lambda lista_tuplas: [" ".join(tupla) for tupla in lista_tuplas]))

len(dictionary_bi_nltk)

dictionary_bi_nltk.save('dictionary_bi_nltk.gensim')

# Crear una copia del diccionario para no modificar el original.
dictionary_filtrado = copy.deepcopy(dictionary_spacy)

# Aplicar filter_extremes
dictionary_filtrado.filter_extremes(no_below=7, no_above=0.7)

dictionary_filtrado.save('dictionary_filtrado_spacy.gensim')

from gensim.corpora import Dictionary, MmCorpus
dictionary = Dictionary.load('dictionary_filtrado_spacy.gensim')

len(dictionary)

# Crear una copia del diccionario para no modificar el original.
dictionary_filtrado_bi = copy.deepcopy(dictionary_bi_spacy)

# Aplicar filter_extremes
dictionary_filtrado_bi.filter_extremes(no_below=7, no_above=0.7)

len(dictionary_filtrado_bi)

dictionary_filtrado_bi.save('dictionary_filtrado_bi.gensim')

# Crear una copia del diccionario para no modificar el original.
dictionary_filtrado_bi_nltk = copy.deepcopy(dictionary_bi_nltk)

# Aplicar filter_extremes
dictionary_filtrado_bi_nltk.filter_extremes(no_below=7, no_above=0.7)

len(dictionary_filtrado_bi_nltk)

## Tablas de frecuencia de palabras

# Obtener el mapeo id -> palabra
token2word = {id_: word for word, id_ in dictionary_filtrado.token2id.items()}

# Obtener la frecuencia de cada palabra
token_frequencies = dictionary_filtrado.dfs  # cfs (collection frequency statistics)

# Convertirlo en un DataFrame ordenado
df_documentos = pd.DataFrame(list(token_frequencies.items()), columns=["token_id", "frequency"])
df_documentos["word"] = df_documentos["token_id"].map(token2word)  # Mapear id a palabra

# Ordenar de mayor a menor frecuencia
df_documentos = df_documentos.sort_values(by="frequency", ascending=False).reset_index(drop=True)

print(df_documentos)

# Obtener el mapeo id -> palabra
token2word = {id_: word for word, id_ in dictionary_filtrado_bi.token2id.items()}

# Obtener la frecuencia de cada palabra
token_frequencies = dictionary_filtrado_bi.dfs  # cfs (collection frequency statistics)

# Convertirlo en un DataFrame ordenado
df_documentos = pd.DataFrame(list(token_frequencies.items()), columns=["token_id", "frequency"])
df_documentos["word"] = df_documentos["token_id"].map(token2word)  # Mapear id a palabra

# Ordenar de mayor a menor frecuencia
df_documentos = df_documentos.sort_values(by="frequency", ascending=False).reset_index(drop=True)

print(df_documentos)

df_documentos[:10]

corpus = [dictionary_filtrado.doc2bow(text) for text in df_preprocesado['texto_preprocesado']]

corpus_bi = [dictionary_filtrado_bi.doc2bow(text) for text in df_preprocesado['bigramas_preprocesado']]

# Guardar el corpus BoW en formato .mm
MmCorpus.serialize('corpus_bi.mm', corpus_bi)

# Analisis de Servicios 

pares_unicos = set(zip(df_preprocesado['service_name'], df_preprocesado['service_code']))

# Mostrar los pares
for par in sorted(pares_unicos, key=lambda x: x[0]):
    print(par)

df_pares = pd.DataFrame(pares_unicos, columns=['columna_1', 'columna_2'])

# Guardar el DataFrame como archivo CSV
df_pares.to_csv('pares_unicos.csv', index=False)

len(df_preprocesado['service_code'].unique())

df_preprocesado = pd.read_csv('df_preprocesado.csv', sep=';')

# Servicios unificados

import pandas as pd
df_service_code = pd.read_csv('servicios_unicos.csv', sep=",",header=None)

df_service_code.head(20)

df_service_code = df_service_code.rename(columns={2: 'service_code_unificado'})

df_service_code[1]

import pandas as pd

def agregar_code_service_unificado(df_preprocesado, df_service_code):
    # Crear un diccionario para mapear service_code a service_code_unificado
    mapeo_codigos = dict(zip(df_service_code['service_code'], df_service_code['service_code_unificado']))

    # Agregar la nueva columna utilizando el mapeo
    df_preprocesado['code_service_unificado'] = df_preprocesado['service_code'].map(mapeo_codigos)

    return df_preprocesado

# Ejemplo de uso:
# Supongamos que tienes tus DataFrames df_preprocesado y df_service_code cargados
# df_preprocesado = pd.read_csv("tu_archivo_preprocesado.csv")
# df_service_code = pd.read_csv("tu_archivo_service_code.csv")

df_preprocesado_con_codigo = agregar_code_service_unificado(df_preprocesado, df_service_code)
# print(df_preprocesado_con_codigo)

len(df_preprocesado['code_service_unificado'].unique())
