==> 03_02_evaluacion_cuantitativa_bigramas.ipynb <==
import pandas as pd
import os
import time
import numpy as np
from gensim.models import LdaModel
import ast
from gsdmm import MovieGroupProcess
from bertopic import BERTopic
from sentence_transformers import SentenceTransformer
from gensim.corpora import Dictionary, MmCorpus
from gensim.models import Word2Vec
from sklearn.cluster import KMeans,AgglomerativeClustering
from sklearn.metrics import adjusted_rand_score, normalized_mutual_info_score, v_measure_score, confusion_matrix, rand_score, fowlkes_mallows_score, davies_bouldin_score, calinski_harabasz_score, silhouette_score
from collections import Counter,defaultdict
from bcubed_metrics.bcubed import Bcubed
from gensim import models
import pickle
import bz2
from sklearn.feature_extraction.text import TfidfVectorizer

## Funciones para tratar datos 

def filter_documents_for_dictionary(tokenized_documents, filtered_words):
    filtered_documents = []
    for doc in tokenized_documents:
        filtered_doc = [word for word in doc if word in filtered_words]
        filtered_documents.append(filtered_doc)
    return filtered_documents

def guardar_objeto(resultados, nombre_archivo):
    with open(nombre_archivo, "wb") as archivo:
        pickle.dump(resultados, archivo)
        
# matriz donde cada fila representa un vector de documento, obtenido como el promedio de los vectores de las palabras que lo componen
def get_document_vectors(model, documents):
    document_vectors = []
    for doc in documents:
        word_vectors = [model.wv[word] for word in doc if word in model.wv.key_to_index]
        if word_vectors:
            document_vector = np.mean(word_vectors, axis=0)
            document_vectors.append(document_vector)
        else:
            # Manejar documentos vacíos o sin palabras válidas
            # Omitir el documento o asignar un vector de ceros
            document_vectors.append(np.zeros(model.vector_size)) #vector de ceros
            pass #omitir el documento
    return document_vectors

## Funciones para evaluar modelos

def create_bcubed_predicted_clustering(labels, true_labels):
    cluster_to_service_counts = defaultdict(lambda: defaultdict(int))
    for i, cluster in enumerate(labels):
        service = true_labels[i]
        cluster_to_service_counts[cluster][service] += 1
    predicted_clustering = []
    for cluster, service_counts in cluster_to_service_counts.items():
        predicted_clustering.append(dict(service_counts))
    return predicted_clustering


def convertir_claves_a_string_en_predicted_clustering(predicted_clustering):
    nuevo_predicted_clustering = []
    for topic_counts in predicted_clustering:
        nuevo_topic_counts = convertir_claves_a_string(topic_counts)
        nuevo_predicted_clustering.append(nuevo_topic_counts)
    return nuevo_predicted_clustering


def contar_codigos_servicio(true_labels):
    conteo_servicios = {}
    for servicio in true_labels:
        if servicio in conteo_servicios:
            conteo_servicios[servicio] += 1
        else:
            conteo_servicios[servicio] = 1
    return conteo_servicios

def convertir_claves_a_string(diccionario):
    nuevo_diccionario = {}
    for clave, valor in diccionario.items():
        nueva_clave = str(clave)  # Convierte la clave a string
        nuevo_diccionario[nueva_clave] = valor
    return nuevo_diccionario


def evaluar_modelo(etiquetas_reales, etiquetas_predichas, X=None):
    
    resultados = {}

    resultados["adjusted_rand_score"] = adjusted_rand_score(
        etiquetas_reales, etiquetas_predichas
    )
    resultados["normalized_mutual_info_score"] = normalized_mutual_info_score(
        etiquetas_reales, etiquetas_predichas
    )
    
    resultados["confusion_matrix"] = confusion_matrix(
        etiquetas_reales, etiquetas_predichas
    )
   
    resultados["fowlkes_mallows_score"] = fowlkes_mallows_score(
        etiquetas_reales, etiquetas_predichas
    )
    
    # Métricas B-Cubed
    #Preparar diccionario para cluster indicar documentos asociados a etiquetas reales
    
    predicted_clustering = create_bcubed_predicted_clustering(etiquetas_predichas , etiquetas_reales)
    predicted_clustering_string = convertir_claves_a_string_en_predicted_clustering(predicted_clustering)
   
    #Preparar diccionario con ground_truth_cluster
    
    conteo_servicios = contar_codigos_servicio(etiquetas_reales)
    ground_labels = convertir_claves_a_string(conteo_servicios)
   
    b = Bcubed(ground_truth_clustering=ground_labels, predicted_clustering=predicted_clustering_string)
   
    resultados["bcubed_precision"] = b.bcubed_precision
    resultados["bcubed_recall"] = b.bcubed_recall
    resultados["bcubed_f1"] = b.bcubed_f1
    
    # Métricas de validación interna
    
    if X is not None:
        resultados["davies_bouldin_score"] = davies_bouldin_score(
          X, etiquetas_predichas
      )
        resultados["calinski_harabasz_score"] = calinski_harabasz_score(
          X, etiquetas_predichas
      )
        resultados["silhouette_score"] = silhouette_score(
          X, etiquetas_predichas
      )
        resultados["homogeneity_score"] = homogeneity_score(
        etiquetas_reales, etiquetas_predichas
    )
        resultados["completeness_score"] = completeness_score(
        etiquetas_reales, etiquetas_predichas
    )
        



    return resultados

## Función principal para generar y evaluar modelos 

def generar_y_evaluar_modelos_topicos(etiquetas_reales, modelos=["lda-bi", "hlda", "gsdmm", "bertopic_Roberta","bertopic", "agglomerative_word2vec","agglomerative_tfidf", "kmeans_Tfidf","kmeans_word2vec_100","kmeans_word2vec_150","kmeans_word2vec_200","kmeans_word2vec_250"], num_ejecuciones=5, **kwargs):
    
    for modelo_nombre in modelos:
        resultados_bi[modelo_nombre] = {}
        resultados_bi[modelo_nombre]["etiquetas_predichas"] = []  # Inicializa la lista de etiquetas predichas
        resultados_ejecuciones = []
        tiempos_entrenamiento = []
        random_seeds = [42, 123, 567, 890, 1001]
        modelo = None
        X = None
        if modelo_nombre == "bertopic" or modelo_nombre == "bertopic_Roberta": 
            for ejecucion in range(num_ejecuciones):
                tiempo_inicio = time.time()
                if modelo_nombre == "bertopic":
                    documentos = kwargs.get("documentos")
                    #text_documents = [" ".join(doc) for doc in documentos]
                    vectorizador = TfidfVectorizer(ngram_range=(2, 2))
                    bertopic_model = BERTopic(vectorizer_model=vectorizador,language="spanish",nr_topics=len(set(etiquetas_reales)),verbose=True)
                    etiquetas_predichas, _ = bertopic_model.fit_transform(documentos)
                    # Obtenemos las representaciones vectoriales de BERTopic
                    # embedding_model = SentenceTransformer("all-mpnet-base-v2")
                    # X = embedding_model.encode(documentos)  
                elif modelo_nombre == "bertopic_Roberta":
                    documentos = kwargs.get("documentos")
                    #text_documents = [" ".join(doc) for doc in documentos]
                    # Cargar un modelo RoBERTa pre-entrenado
                    embedding_model = SentenceTransformer("roberta-base-nli-stsb-mean-tokens")
                    bertopic_model = BERTopic(nr_topics=len(set(etiquetas_reales)),verbose=True,embedding_model=embedding_model)
                    etiquetas_predichas, _ = bertopic_model.fit_transform(documentos)
                tiempo_fin = time.time()
                tiempos_entrenamiento.append(tiempo_fin - tiempo_inicio)
                if 'bertopic_model' in locals(): 
                    modelo = bertopic_model
                    nombre_archivo = os.path.join(ruta_guardado, f"{modelo_nombre}_ejecucion_{ejecucion}.pkl") # Cambia la extensión a .pkl
                    with open(nombre_archivo, "wb") as archivo:
                        pickle.dump(modelo, archivo)
                    resultados_ejecuciones.append(evaluar_modelo(etiquetas_reales, etiquetas_predichas, X))
                    resultados_bi[modelo_nombre]["resultados_ejecuciones"] = resultados_ejecuciones

        if modelo_nombre == "agglomerative_word2vec":
            documentos = kwargs.get("documentos")
            agg_model = AgglomerativeClustering(n_clusters=len(set(etiquetas_reales)), linkage='ward')
            etiquetas_predichas = agg_model.fit_predict(documentos)
        elif modelo_nombre == "agglomerative_Tfidf":
            tiempo_inicio = time.time()
            documentos = kwargs.get("documentos")
            vectorizer = TfidfVectorizer()
            documentos_como_cadenas = [" ".join(doc) for doc in documentos]
            # Calcular TF-IDF
            tfidf_matrix = vectorizer.fit_transform(documentos_como_cadenas)
            tfidf_matrix_dense = tfidf_matrix.toarray()
            agg_model = AgglomerativeClustering(n_clusters=len(set(etiquetas_reales)), linkage='ward')
            etiquetas_predichas = agg_model.fit_predict(tfidf_matrix_dense)
            tiempo_fin = time.time()
            tiempos_entrenamiento.append(tiempo_fin - tiempo_inicio)
        if 'agg_model' in locals(): 
            modelo = agg_model
            nombre_archivo = os.path.join(ruta_guardado, f"{modelo_nombre}_ejecucion_0.pkl") # Cambia la extensión a .pkl
            with open(nombre_archivo, "wb") as archivo:
                pickle.dump(modelo, archivo)
                    
            resultados_bi[modelo_nombre]["etiquetas_predichas"].append(etiquetas_predichas)
            resultados_ejecuciones.append(evaluar_modelo(etiquetas_reales, etiquetas_predichas, X))
            resultados_bi[modelo_nombre]["resultados_ejecuciones"] = resultados_ejecuciones
            
        for seed in random_seeds:
            tiempo_inicio = time.time()
           

            if modelo_nombre == "lda-bi":
                corpus = kwargs.get("corpus")
                diccionario = kwargs.get("diccionario")
                if corpus is None:
                    raise ValueError("corpus debe proporcionarse para el modelo LDA.")
                if diccionario is None:
                    raise ValueError("diccionario debe proporcionarse para el modelo GSDMM.")                
                lda_model = models.LdaModel(corpus, num_topics=len(set(etiquetas_reales)), id2word=diccionario, alpha=0.05, eta=0.01, iterations=100, random_state=seed  )
                etiquetas_predichas = [max(lda_model[doc], key=lambda x: x[1])[0] for doc in corpus]
                
                
                          
                
            elif modelo_nombre == "gsdmm":
                n_terms = kwargs.get("n_terms")
                documentos = kwargs.get("documentos")
                if n_terms is None:
                    raise ValueError("n_terms debe proporcionarse para el modelo GSDMM.")
                if documentos is None:
                    raise ValueError("documentos debe proporcionarse para el modelo GSDMM.")    
                gsdmm_model = MovieGroupProcess(K=len(set(etiquetas_reales)), n_iters=30, alpha=0.1, beta=0.1)
                y = gsdmm_model.fit(documentos,n_terms)
                etiquetas_predichas = np.array(y)
                
         
          
                            
            elif modelo_nombre == "kmeans_Tfidf":
                documentos = kwargs.get("documentos")
                vectorizador = TfidfVectorizer(ngram_range=(2, 2))
                documentos_como_cadenas = [" ".join(doc) for doc in documentos]
                # Calcular TF-IDF
                tfidf_matrix = vectorizador.fit_transform(documentos_como_cadenas)
                kmeans_model = KMeans(n_clusters=len(set(etiquetas_reales)), max_iter=100,random_state=seed )
                etiquetas_predichas = kmeans_model.fit_predict(tfidf_matrix)
            elif modelo_nombre == "kmeans_word2vec" or modelo_nombre == "kmeans_word2vec_100" or modelo_nombre == "kmeans_word2vec_150" or modelo_nombre == "kmeans_word2vec_200" or modelo_nombre == "kmeans_word2vec_250":
                documentos = kwargs.get("documentos")
                kmeans_model = KMeans(n_clusters=len(set(etiquetas_reales)), max_iter=100, random_state=seed )
                etiquetas_predichas = kmeans_model.fit_predict(documentos)                

            tiempo_fin = time.time()
            tiempos_entrenamiento.append(tiempo_fin - tiempo_inicio)
            
        # Guardar el modelo
            if 'lda_model' in locals(): modelo = lda_model
            elif 'hlda_model' in locals(): modelo = hlda_model
            elif 'gsdmm_model' in locals(): modelo = gsdmm_model
            elif 'kmeans_model' in locals(): modelo = kmeans_model
                 
            # Guardar el modelo
            if modelo is not None:
                nombre_archivo = os.path.join(ruta_guardado, f"{modelo_nombre}_ejecucion_{seed}.pkl") # Cambia la extensión a .pkl
                with open(nombre_archivo, "wb") as archivo:
                    pickle.dump(modelo, archivo)
                    
            resultados_bi[modelo_nombre]["etiquetas_predichas"].append(etiquetas_predichas)
                
            resultados_ejecuciones.append(evaluar_modelo(etiquetas_reales, etiquetas_predichas, X))
            resultados_bi[modelo_nombre]["resultados_ejecuciones"] = resultados_ejecuciones
            
    # Promediar y calcular la desviación estándar de los resultados
        for metrica in resultados_ejecuciones[0]:
            valores_metrica = [resultado[metrica] for resultado in resultados_ejecuciones]
            resultados_bi[modelo_nombre][metrica] = {
                "promedio": np.mean(valores_metrica),
                "desviacion_estandar": np.std(valores_metrica)
            }

        resultados_bi[modelo_nombre]["tiempo_entrenamiento"] = {
            "promedio": np.mean(tiempos_entrenamiento),
            "desviacion_estandar": np.std(tiempos_entrenamiento)
        }
    guardar_objeto(resultados_bi, "resultados_modelos_bi.pkl")   
    return resultados_bi

### Función para visualizar resultados 

def generar_tabla_metricas_fijas2(metricas, metricas_deseadas, modelos_deseados=None):
    tabla_html = "<table>\n"
    tabla_html += "  <tr>\n"
    tabla_html += "    <th>Métrica</th>\n"

    # Determinar qué modelos mostrar
    modelos_a_mostrar = modelos_deseados if modelos_deseados else list(metricas.keys())

    for modelo in modelos_a_mostrar:
        tabla_html += f"    <th>{modelo.upper()}</th>\n"
    tabla_html += "  </tr>\n"

    for metrica in metricas_deseadas:
        tabla_html += "  <tr>\n"
        tabla_html += f"    <td>{metrica.replace('_', ' ').title()}</td>\n"
        for modelo in modelos_a_mostrar:
            if metrica in metricas[modelo]:
                promedio = metricas[modelo][metrica]['promedio']
                desviacion = metricas[modelo][metrica]['desviacion_estandar']
                tabla_html += f"    <td>{promedio:.4f} ± {desviacion:.4f}</td>\n"
            else:
                tabla_html += "    <td>-</td>\n"  # Si la métrica no está para un modelo
        tabla_html += "  </tr>\n"

    tabla_html += "</table>"
    return tabla_html

# Ejemplo de uso con modelos específicos:
metricas_deseadas = [
    'adjusted_rand_score',
    'normalized_mutual_info_score',
    'fowlkes_mallows_score',
    'bcubed_precision',
    'bcubed_recall',
    'bcubed_f1',
    'tiempo_entrenamiento'
]



# carga de datos

#Cargar datos, diccionario y corpus preparados en preprocesado 

# df_preprocesado = pd.read_csv('df_preprocesado.csv', sep=';')
dictionary = Dictionary.load('recursos/dictionary_filtrado_bi.gensim')
corpus = MmCorpus('recursos/corpus_bi.mm')


#Cargar etiquetas de servicios asociados a documentos

#true_labels = df_preprocesado['code_service_unificado'].tolist()
with open('recursos/true_labeles.pkl', 'rb') as archivo:
    true_labels = pickle.load(archivo)

#Iniciarlizar resultados y documentos

documentos=[]
resultados_bi = {}

#Preparar ruta para guardado

ruta_guardado = "./modelos_topicos_bi"
if not os.path.exists(ruta_guardado):
    os.makedirs(ruta_guardado)
 

#preparar documentos para partir de las mismas condiciones 
filtered_words = set(dictionary.token2id.keys())
n_term=len(filtered_words)
#preprocesado_documents=df_preprocesado['texto_preprocesado'].apply(ast.literal_eval)    
#documents = filter_documents_for_dictionary(preprocesado_documents, filtered_words)
with open('recursos/documents_bi.pkl', 'rb') as archivo:
    documents = pickle.load(archivo)

## Generar y evaluar modelos 

### LDA

resultados_bi = generar_y_evaluar_modelos_topicos(true_labels, modelos=["lda-bi"],diccionario=dictionary, corpus=corpus)      

modelos_a_mostrar = ['lda-bi'] # Solo muestra estos modelos.

tabla_html = generar_tabla_metricas_fijas2(resultados_bi, metricas_deseadas, modelos_a_mostrar)

from IPython.display import display, HTML
display(HTML(tabla_html))

### GSDMM

with open('recursos/resultados_modelos_bi.pkl', 'rb') as archivo:
    resultados = pickle.load(archivo)

resultados = generar_y_evaluar_modelos_topicos(true_labels, modelos=["gsdmm"],n_terms=n_term,documentos=documents)

modelos_a_mostrar = ['gsdmm'] # Solo muestra estos modelos.
tabla5_html = generar_tabla_metricas_fijas2(resultados, metricas_deseadas, modelos_a_mostrar)
display(HTML(tabla5_html))

### BERTOPIC

with open('recursos/resultados_modelos_bi.pkl', 'rb') as archivo:
    resultados = pickle.load(archivo)

#Si utilizamos documentos originales
# documents_original=df_preprocesado['texto']
# documents_original
with bz2.BZ2File('recursos/documentos_originales.pkl.bz2', 'rb') as f:
    documents_original = pickle.load(f)

resultados = generar_y_evaluar_modelos_topicos(true_labels, modelos=["bertopic"],documentos=documents_original)

modelos_a_mostrar = ['bertopic'] # Solo muestra estos modelos.

tabla_html2 = generar_tabla_metricas_fijas2(resultados, metricas_deseadas, modelos_a_mostrar)
display(HTML(tabla_html2))

### Kmeans_word2vec

with open('resultados_modelos.pkl', 'rb') as archivo:
    resultados = pickle.load(archivo)

#### Generar modelos partiendo de Word2Vec entre 100 y 250 epocs

epochs_to_test = [100, 150, 200, 250]  # Números de épocas a probar
all_document_vectors_bi = {}  # Diccionario para almacenar los vectores de documentos

for epochs in epochs_to_test:
    model_name = f"w2v_{epochs}_epochs"
    model = Word2Vec(
        sentences=documents, vector_size=100, window=5, min_count=1, workers=4, seed=42
    )
    model.train(documents, total_examples=len(documents), epochs=epochs)
    model.save(f"w2v_{epochs}_epochs.model") #guardamos el modelo.
    document_vectors = get_document_vectors(model, documents)
    document_vectors = [vector for vector in document_vectors if vector is not None]
    document_vectors = np.array(document_vectors)
    all_document_vectors_bi[model_name] = document_vectors  # Guardamos los vectores de documentos.

len(all_document_vectors_bi["w2v_100_epochs"])

with open('all_document_vectors_bi.pkl', "wb") as archivo:
                    pickle.dump(all_document_vectors_bi, archivo)

resultados_word2vec={}

resultados_word2vec["kmeans_word2vec_100"] = generar_y_evaluar_modelos_topicos(true_labels, modelos=["kmeans_word2vec_100"],documentos=all_document_vectors_bi["w2v_100_epochs"])

resultados = generar_y_evaluar_modelos_topicos(true_labels, modelos=["kmeans_word2vec_150"],documentos=all_document_vectors_bi["w2v_150_epochs"])

resultados = generar_y_evaluar_modelos_topicos(true_labels, modelos=["kmeans_word2vec_200"],documentos=all_document_vectors_bi["w2v_200_epochs"])

resultados = generar_y_evaluar_modelos_topicos(true_labels, modelos=["kmeans_word2vec_250"],documentos=all_document_vectors_bi["w2v_250_epochs"])

from IPython.display import display, HTML
modelos_a_mostrar = ['kmeans_word2vec_100','kmeans_word2vec_150','kmeans_word2vec_200','kmeans_word2vec_250'] # Solo muestra estos modelos.

tabla_html = generar_tabla_metricas_fijas2(resultados_bi, metricas_deseadas, modelos_a_mostrar)
display(HTML(tabla_html))

### KMEANS tfidf

resultados = generar_y_evaluar_modelos_topicos(true_labels, modelos=["kmeans_Tfidf"],documentos=documents)

modelos_a_mostrar = ['kmeans_Tfidf'] # Solo muestra estos modelos.

tabla_html = generar_tabla_metricas_fijas2(resultados, metricas_deseadas, modelos_a_mostrar)

from IPython.display import display, HTML
display(HTML(tabla_html))

### Aglomerative

with open('recursos/resultados_modelos_bi.pkl', 'rb') as archivo:
    resultados_bi = pickle.load(archivo)

with open('all_document_vectors_bi.pkl', 'rb') as archivo:
    all_document_vectors = pickle.load(archivo)

resultados = generar_y_evaluar_modelos_topicos(true_labels, modelos=["agglomerative_word2vec"],documentos=all_document_vectors_bi["w2v_250_epochs"])

resultados = generar_y_evaluar_modelos_topicos(true_labels, modelos=["agglomerative_Tfidf"],documentos=documents)
